+++ gcl/o/reference.c
@@ -74,7 +74,7 @@ Lsymbol_function()
 	if (type_of(sym) != t_symbol)
 		not_a_symbol(sym);
 	if (sym->s.s_sfdef != NOT_SPECIAL) {
-		vs_push(make_fixnum((int)(sym->s.s_sfdef)));
+		vs_push(make_fixnum((long)(sym->s.s_sfdef)));
 		vs_base[0] = sLspecial;
 		stack_cons();
 		return;
